M-02 : Push 59 : ex01 et ex02 : passage de round a roundf

+++ module-02/ex01/Fixed.cpp
@@ -6,7 +6,7 @@
 /*   By: guderram <guderram@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2022/09/04 08:40:09 by guderram          #+#    #+#             */
-/*   Updated: 2023/03/22 23:51:39 by guderram         ###   ########.fr       */
+/*   Updated: 2023/05/15 11:54:39 by guderram         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -22,13 +22,13 @@ Fixed::Fixed(void) : _val(0)
 }
 
 
-Fixed::Fixed(const int val) : _val((int)round(val * (1 << this->_bit)))
+Fixed::Fixed(const int val) : _val((int)roundf(val * (1 << this->_bit)))
 {
 	std::cout << "Int constructor called" << std::endl;
 	return ;
 }
 
-Fixed::Fixed(const float val) : _val((int)round(val * (1 << this->_bit)))
+Fixed::Fixed(const float val) : _val((int)roundf(val * (1 << this->_bit)))
 {
 	std::cout << "Float constructor called" << std::endl;
 	return ;
